@@ -8,9 +8,61 @@
   </div>
 {/snippet}
 
+{#snippet UIMessageToolResult(part)}
+  {#if part.status === 'error' && part.error}
+    <pre>
+      {String(part.error)}
+    </pre>
+  {/if}
+
+  {#if part.status === 'complete' && part.result}
+    {@const result = part.result}
+    {#if typeof result === 'string'}
+      <pre>{result}</pre>
+    {/if}
+
+    {#if Array.isArray(result)}
+      <div>
+        {#each result as item}
+          <li>
+            {#if item.type === 'text'}
+              <div>{String(item)}</div>
+            {/if}
+            {#if item.type === 'image_url'}
+              <img src={String(item.image_url)} alt='Tool Result' style='max-width: 100%; border-radius: 4px;' />
+            {/if}
+            {#if item.type === 'input_audio'}
+              <audio controls>
+                <source src={item.input_audio.data} type={`audio/${item.input_audio.format}`} />
+                Your browser does not support the audio element.
+              </audio>
+            {/if}
+          </li>
+        {/each}
+      </div>
+    {/if}
+  {/if}
+
+{/snippet}
+
 {#snippet UIMessageToolPart(part)}
-  <div>
-    {part.toolCall.name}
+  {@const hasResult = part.status === 'complete' || part.status === 'error'}
+  {@const isLoading = part.status === 'loading' || part.status === 'partial'}
+  <div
+    class="collapse collapse-arrow border {part.status === 'error' ? 'bg-red-100' : 'bg-base-100'}"
+  >
+    <input type='checkbox' class='tweak-collapse' />
+    <div class='collapse-title font-semibold tweak-collapse tweak-collapse-title-arrow'>{part.toolCall.function.name}</div>
+    <div class='collapse-content'>
+      {#if isLoading}
+        <div class='skeleton h-4 w-full'></div>
+      {/if}
+      {#if hasResult}
+        <div>
+          {@render UIMessageToolResult(part)}
+        </div>
+      {/if}
+    </div>
   </div>
 {/snippet}
 
@@ -30,12 +82,9 @@
     class={[
       'chat-bubble',
       message.role === 'user' ? 'chat-bubble-primary' : '',
-      'chatBubble',
     ]}
   >
-
     {#each message.parts as part, index (index)}
-
       {#if part.type === 'text'}
         {@render UIMessageTextPart(part)}
       {:else if part.type === 'tool-call'}
@@ -45,7 +94,7 @@
       {/if}
     {/each}
     {#if isError}
-      <div class='errorMessage'>
+      <div class='error-message'>
         ❌
         {error?.message}
       </div>
@@ -59,13 +108,23 @@
 </div>
 
 <style>
-.chatBubble {
+.chat-bubble {
   display: flex;
   flex-direction: column;
   gap: .5rem;
 }
 
-.errorMessage {
+.error-message {
   font-size: 12px;
+}
+
+.tweak-collapse {
+  padding-top: .5rem;
+  padding-bottom: .5rem;
+  min-height: 2.75rem;
+}
+
+.tweak-collapse-title-arrow:after {
+  top: 1.4rem;
 }
 </style>
